requirements.txt
torch
pyspark
pandas 
dvc
mlflow==2.2.2
notebook
numpy
matplotlib
seaborn
python-box
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
scipy
Flask
Flask-Cors
gdown
torchvision
tensorflow
keras
idx2numpy
mlflow==2.2.2
numpy==1.26.4
scikit-learn==1.2.2
pyarrow==11.0.0
pandas==2.2.2
matplotlib==3.8.4
seaborn==0.13.2
-e .